requirements.txt
Django==1.9.6
Pillow==3.2.0
python-resize-image==1.1.3
